test(reservation-domain): TimeRange, SeatLocator Fixture 추가
